Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
google-generativeai==0.3.2
numpy==1.24.3
requests==2.31.0
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.3